task delete done

+++ src/router/taskRouter.js
@@ -54,7 +54,7 @@ router.post("/tasks",auth, async (req, res) => {
     }
   });
   
-  ///UPDATE-PATCH///
+  ///UPDATE ONLY AUTH TASK-PATCH///
   
   router.patch("/tasks/:id", auth,  async (req, res) => {
     const updates = Object.keys(req.body);
@@ -84,11 +84,11 @@ router.post("/tasks",auth, async (req, res) => {
     }
   });
   
-  ///DELETE-DELETE///
+  ///DELETE ONLY AUTH TASK-DELETE///
   
-  router.delete('/tasks/:id', async (req,res) => {
+  router.delete('/tasks/:id',auth, async (req,res) => {
       try {
-          const task = await Task.findByIdAndDelete(req.params.id);
+          const task = await Task.findByIdAndDelete({_id:req.params.id,owner: req.user._id});
           if(!task){
               return res.status(404).send()
           }
